refactor: AuthenticationPrincipalArgumentResolver 의 Member 의존성 제거
