+++ src/main/java/nextstep/member/application/service/MemberService.java
@@ -1,9 +1,10 @@
 package nextstep.member.application.service;
 
 import nextstep.auth.application.service.UserDetailService;
+import nextstep.auth.domain.UserDetail;
 import nextstep.member.application.dto.MemberRequest;
 import nextstep.member.application.dto.MemberResponse;
-import nextstep.member.domain.LoginMember;
+import nextstep.auth.domain.LoginMember;
 import nextstep.member.domain.entity.Member;
 import nextstep.member.domain.repository.MemberRepository;
 import org.springframework.stereotype.Service;
@@ -12,7 +13,7 @@ import javax.persistence.EntityNotFoundException;
 
 @Service
 public class MemberService implements UserDetailService {
-    private MemberRepository memberRepository;
+    private final MemberRepository memberRepository;
 
     public MemberService(MemberRepository memberRepository) {
         this.memberRepository = memberRepository;
